User Model and Integration Test
Added the mysql jdbc driver.
Created the User Domain Class with properties userId, password, motto,
dateCreated, and lastUpdated.
Created an Integration test for the User Domain Class that deletes all
users on setup and then adds a new user to the database and asserts that
user.count is 1, i.e. the insert was successful.